requirements.txt
streamlit
pandas
python-docx
PyMuPDF
sentence-transformers==2.2.2
torch==2.0.1
torchvision==0.15.2
tqdm
transformers==4.26.1
huggingface-hub==0.13.4